Split records into their own test
