+++ crates/test-rust-wasm/src/bin/records.rs
@@ -1,0 +1,3 @@
+include!("../../../../tests/runtime/records/wasm.rs");
+
+fn main() {}
