+++ tests/host.witx
@@ -1,56 +1,3 @@
-// ===========================================
-// scalars
-// ===========================================
-
-// roundtrip_usize: function(a: usize) -> usize
-multiple_results: function() -> (u8, u16)
-
-// ===========================================
-// records
-// ===========================================
-
-swap_tuple: function(a: tuple<u8, u32>) -> tuple<u32, u8>
-
-flags f1 { a, b }
-roundtrip_flags1: function(a: f1) -> f1
-
-flags f2 { c, d, e }
-roundtrip_flags2: function(a: f2) -> f2
-
-flags flag8 {
-  b0, b1, b2, b3, b4, b5, b6, b7,
-}
-
-flags flag16 {
-  b0, b1, b2, b3, b4, b5, b6, b7,
-  b8, b9, b10, b11, b12, b13, b14, b15,
-}
-
-flags flag32 {
-  b0, b1, b2, b3, b4, b5, b6, b7,
-  b8, b9, b10, b11, b12, b13, b14, b15,
-  b16, b17, b18, b19, b20, b21, b22, b23,
-  b24, b25, b26, b27, b28, b29, b30, b31,
-}
-
-flags flag64 {
-  b0, b1, b2, b3, b4, b5, b6, b7,
-  b8, b9, b10, b11, b12, b13, b14, b15,
-  b16, b17, b18, b19, b20, b21, b22, b23,
-  b24, b25, b26, b27, b28, b29, b30, b31,
-  b32, b33, b34, b35, b36, b37, b38, b39,
-  b40, b41, b42, b43, b44, b45, b46, b47,
-  b48, b49, b50, b51, b52, b53, b54, b55,
-  b56, b57, b58, b59, b60, b61, b62, b63,
-}
-
-roundtrip_flags3: function(a: flag8, b: flag16, c: flag32, d: flag64) -> (flag8, flag16, flag32, flag64)
-
-record r1 { a: u8, b: f1 }
-roundtrip_record1: function(a: r1) -> r1
-
-tuple0: function(a: tuple<>) -> tuple<>
-tuple1: function(a: tuple<u8>) -> tuple<u8>
 
 // ===========================================
 // variants
