+++ tests/runtime/records/exports.witx
@@ -1,0 +1,46 @@
+test_imports: function()
+
+multiple_results: function() -> (u8, u16)
+
+swap_tuple: function(a: tuple<u8, u32>) -> tuple<u32, u8>
+
+flags f1 { a, b }
+roundtrip_flags1: function(a: f1) -> f1
+
+flags f2 { c, d, e }
+roundtrip_flags2: function(a: f2) -> f2
+
+flags f8 {
+  b0, b1, b2, b3, b4, b5, b6, b7,
+}
+
+flags f16 {
+  b0, b1, b2, b3, b4, b5, b6, b7,
+  b8, b9, b10, b11, b12, b13, b14, b15,
+}
+
+flags "f32" {
+  b0, b1, b2, b3, b4, b5, b6, b7,
+  b8, b9, b10, b11, b12, b13, b14, b15,
+  b16, b17, b18, b19, b20, b21, b22, b23,
+  b24, b25, b26, b27, b28, b29, b30, b31,
+}
+
+flags "f64" {
+  b0, b1, b2, b3, b4, b5, b6, b7,
+  b8, b9, b10, b11, b12, b13, b14, b15,
+  b16, b17, b18, b19, b20, b21, b22, b23,
+  b24, b25, b26, b27, b28, b29, b30, b31,
+  b32, b33, b34, b35, b36, b37, b38, b39,
+  b40, b41, b42, b43, b44, b45, b46, b47,
+  b48, b49, b50, b51, b52, b53, b54, b55,
+  b56, b57, b58, b59, b60, b61, b62, b63,
+}
+
+roundtrip_flags3: function(a: f8, b: f16, c: "f32", d: "f64") -> (f8, f16, "f32", "f64")
+
+record r1 { a: u8, b: f1 }
+roundtrip_record1: function(a: r1) -> r1
+
+tuple0: function(a: tuple<>) -> tuple<>
+tuple1: function(a: tuple<u8>) -> tuple<u8>
